chore: address review changes
